fix for authentication failure on RHEL9 when sasldb file is not found
The original https://github.com/moriyoshi/cyrus-sasl-xoauth2 stored
the JWT in the oauth2BearerTokens property of the sasldb on the server
and compared it with the JWT presented by the client.
However, cyrus-sasl-xoauth2-idp uses Keycloak to verify the JWT,
so there is no need to verify the sasldb file in the first place.

On Linux systems other than RHEL9 variants, Berkeley DB is used as
the default sasldb format, so even if the sasldb file does not exist,
the authentication succeeded without a fatal error.
However, in RHEL9 variants, Berkeley DB was removed and gdbm became
the default sasldb format, and if the sasldb file did not exist,
a fatal error occurred and authentication failed.